ai-detector-backend-setup/backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
aiosmtplib==3.0.1

# AI/ML
tensorflow==2.15.0
numpy==1.26.3
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
